Sends additional text to contact when the user is 1 minute away

+++ app/src/main/java/com/neilvohra/asdghowns/ridealert/Tasks/RemainingTravelTimeTaskCallback.java
@@ -1,5 +1,5 @@
 package com.neilvohra.asdghowns.ridealert.Tasks;
 
 public interface RemainingTravelTimeTaskCallback {
-    void onTaskCompletion(final boolean success, int remainingTravelTime, int activeAlertsIndex);
+    void onTaskCompletion(final boolean success, double remainingTravelTime, int activeAlertsIndex);
 }
